refactor(rendering): split shader param setters out of shadereffect.cpp
